+++ examples/sparse_tensor_basic.py
@@ -24,29 +24,32 @@
 import torch
 import MinkowskiEngine as ME
 
+data_batch_0 = [
+    [0, 0, 2.1, 0, 0],  #
+    [0, 1, 1.4, 3, 0],  #
+    [0, 0, 4.0, 0, 0]
+]
 
-def data_generation():
-    """
-    Return torch tensors for coordinates and features
-    """
-    data = [[0, 0, 2.1, 0, 0],
-            [0, 1, 1.4, 3, 0],
-            [0, 0, 4,   0, 0]]
-    # coordinates and corresponding features
+data_batch_1 = [
+    [1, 0, 0],  #
+    [0, 2, 0],  #
+    [0, 0, 3]
+]
+
+
+def to_sparse_coo(data):
+    # An intuitive way to extract coordinates and features
     coords, feats = [], []
     for i, row in enumerate(data):
         for j, val in enumerate(row):
             if val != 0:
                 coords.append([i, j])
                 feats.append([val])
-    coords = torch.IntTensor(coords)
-    feats = torch.FloatTensor(feats)
-
-    return coords, feats
+    return torch.IntTensor(coords), torch.FloatTensor(feats)
 
 
 def sparse_tensor_initialization():
-    coords, feats = data_generation()
+    coords, feats = to_sparse_coo(data_batch_0)
     # collate sparse tensor data to augment batch indices
     # Note that it is wrapped inside a list!!
     coords, feats = ME.utils.sparse_collate(coords=[coords], feats=[feats])
@@ -54,13 +57,15 @@ def sparse_tensor_initialization():
 
 
 def sparse_tensor_arithmetics():
-    coords, feats = data_generation()
-    new_coords = torch.IntTensor([[0, 1], [2, 3], [4, 5]])
-    new_feats = torch.rand(len(new_coords), feats.size(1))
+    coords0, feats0 = to_sparse_coo(data_batch_0)
+    coords0, feats0 = ME.utils.sparse_collate(coords=[coords0], feats=[feats0])
+
+    coords1, feats1 = to_sparse_coo(data_batch_1)
+    coords1, feats1 = ME.utils.sparse_collate(coords=[coords1], feats=[feats1])
 
     # sparse tensors
-    A = ME.SparseTensor(coords=coords, feats=feats)
-    B = ME.SparseTensor(coords=new_coords, feats=new_feats)
+    A = ME.SparseTensor(coords=coords0, feats=feats0)
+    B = ME.SparseTensor(coords=coords1, feats=feats1)
 
     # The following fails
     try:
@@ -69,8 +74,8 @@ def sparse_tensor_arithmetics():
         pass
 
     B = ME.SparseTensor(
-        coords=new_coords,
-        feats=new_feats,
+        coords=coords1,
+        feats=feats1,
         coords_manager=A.coords_man,  # must share the same coordinate manager
         force_creation=True  # must force creation since tensor stride [1] exists
     )
@@ -84,7 +89,7 @@ def sparse_tensor_arithmetics():
     # Note that it requires the same coords_key (no need to feed coords)
     D = ME.SparseTensor(
         # coords=coords,  not required
-        feats=feats,
+        feats=feats0,
         coords_manager=A.coords_man,  # must share the same coordinate manager
         coords_key=A.coords_key  # For inplace, must share the same coords key
     )
@@ -104,16 +109,18 @@ def operation_mode():
         ME.SparseTensorOperationMode.SHARE_COORDS_MANAGER)
     print(ME.sparse_tensor_operation_mode())
 
-    coords, feats = data_generation()
-    new_coords = torch.IntTensor([[0, 1], [2, 3], [4, 5]])
-    new_feats = torch.rand(len(new_coords), feats.size(1))
+    coords0, feats0 = to_sparse_coo(data_batch_0)
+    coords0, feats0 = ME.utils.sparse_collate(coords=[coords0], feats=[feats0])
+
+    coords1, feats1 = to_sparse_coo(data_batch_1)
+    coords1, feats1 = ME.utils.sparse_collate(coords=[coords1], feats=[feats1])
 
     for _ in range(2):
         # sparse tensors
-        A = ME.SparseTensor(coords=coords, feats=feats)
+        A = ME.SparseTensor(coords=coords0, feats=feats0)
         B = ME.SparseTensor(
-            coords=new_coords,
-            feats=new_feats,
+            coords=coords1,
+            feats=feats1,
             # coords_manager=A.coords_man,  No need to feed the coords_man
             force_creation=True)
 
@@ -123,7 +130,31 @@ def operation_mode():
         ME.clear_global_coords_man()
 
 
+def decomposition():
+    coords0, feats0 = to_sparse_coo(data_batch_0)
+    coords1, feats1 = to_sparse_coo(data_batch_1)
+    coords, feats = ME.utils.sparse_collate(
+        coords=[coords0, coords1], feats=[feats0, feats1])
+
+    # sparse tensors
+    A = ME.SparseTensor(coords=coords, feats=feats)
+    conv = ME.MinkowskiConvolution(
+        in_channels=1, out_channels=2, kernel_size=3, stride=2, dimension=2)
+    B = conv(A)
+
+    # Extract features and coordinates per batch index
+    coords = B.decomposed_coordinates
+    feats = B.decomposed_features
+    coords, feats = B.decomposed_coordinates_and_features
+
+    # To specify a batch index
+    batch_index = 1
+    coords = B.coordinates_at(batch_index)
+    feats = B.features_at(batch_index)
+
+
 if __name__ == '__main__':
     sparse_tensor_initialization()
     sparse_tensor_arithmetics()
     operation_mode()
+    decomposition()